% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDataset.R
\name{analyze_dataset}
\alias{analyze_dataset}
\title{analyze_dataset}
\usage{
analyze_dataset(
  expObj,
  expObjGroups = NULL,
  fdrCut = 0.05,
  subset = 0,
  repN = 2,
  useAllSamplesAsNegativeControl = FALSE
)
}
\arguments{
\item{expObj}{RangedSummarizedExperiment object.}

\item{expObjGroups}{sample groups. Should be a vectgor of 0 and 1. 0 as control samples.}
}
\value{
Figures and a list of result data.
}
\description{
A function analyze data set
}
\examples{
1
}
