% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est_power}
\alias{est_power}
\title{est_power}
\usage{
est_power(n, w = 1, rho = 2, lambda0 = 5, phi0 = 1, alpha = 0.05, f,
  m = 20000, m1 = 200)
}
\arguments{
\item{n}{Numer of samples.}

\item{w}{Ratio of normalization factors between two groups.}

\item{rho}{minimum fold changes for prognostic genes between two groups.}

\item{lambda0}{Average read counts for prognostic genes.}

\item{phi0}{Dispersion for prognostic genes.}

\item{alpha}{alpha level.}

\item{f}{FDR level}

\item{m}{Total number of genes for testing.}

\item{m1}{Expected number of prognostic genes.}
}
\value{
Estimate power
}
\description{
A function to estitamete the power for differential expression analysis of RNA-seq data.
}
\examples{
n<-63;rho<-2;lambda0<-5;phi0<-0.5;f<-0.01
est_power(n=n, rho=rho, lambda0=lambda0, phi0=phi0,f=f)
}

