% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionTableVis.R
\name{plot_gene_counts_range}
\alias{plot_gene_counts_range}
\title{plot_gene_counts_range}
\usage{
plot_gene_counts_range(expObj, targetSize = NULL)
}
\arguments{
\item{expObj}{RangedSummarizedExperiment object or an expression matrix.}

\item{targetSize}{The target library size to scale to. Will not do scale if set as NULL.}
}
\value{
A barplot.
}
\description{
A function to plot propotion of genes in different count range.
}
\examples{
1
}
