% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionTableVis.R
\name{plot_mappedReads_percent}
\alias{plot_mappedReads_percent}
\title{plot_mappedReads_percent}
\usage{
plot_mappedReads_percent(expObj, groupVar = NULL)
}
\arguments{
\item{expObj}{RangedSummarizedExperiment object generated by recount package.}

\item{groupVar}{variable name in colData(expObj) to be used to group the samples to make boxplot.}
}
\value{
A barplot or boxplot.
}
\description{
A function to plot percent of mapped reads in total reads. Only RangedSummarizedExperiment 
object generated by recount package have total reads information to to this.
}
\examples{
1
}
